refactor: move custom selector to own file
