+++ src/common/hooks/useAppDispatch.ts
@@ -1,9 +1,8 @@
-import {TypedUseSelectorHook, useDispatch, useSelector} from "react-redux";
 import {ActionsTypes} from "@/app/action-creators-types.ts";
 import {AppRootStateType} from "@/app/store.ts";
 import {ThunkDispatch} from "redux-thunk";
+import {useDispatch} from "react-redux";
 
 export type AppDispatchType = ThunkDispatch<AppRootStateType, any, ActionsTypes>;
 
-export const useAppSelector: TypedUseSelectorHook<AppRootStateType> = useSelector;
 export const useAppDispatch = () => useDispatch<AppDispatchType>();
