+++ src/common/hooks/useAppSelector.ts
@@ -1,0 +1,4 @@
+import {TypedUseSelectorHook, useSelector} from "react-redux";
+import {AppRootStateType} from "@/app/store.ts";
+
+export const useAppSelector: TypedUseSelectorHook<AppRootStateType> = useSelector;
