+++ src/features/profile/components/PostsBlock/Posts/Posts.tsx
@@ -1,10 +1,9 @@
 import cls from "@/features/profile/components/PostsBlock/Posts/Posts.module.scss";
 import {Post} from "@/features/profile/components/PostsBlock/Posts/Post/Post.tsx";
 import {postsSelector} from "@/features/profile/selectors/postsSelector";
-import {useAppSelector} from "@/common/hooks/useAppDispatch.ts";
+import {useAppSelector} from "@/common/hooks/useAppSelector.ts";
 
 export const Posts = () => {
-
    const postsData = useAppSelector(postsSelector);
 
    const postsList = postsData.length
