+++ src/features/profile/components/PostsBlock/PostsBlock.tsx
@@ -2,7 +2,8 @@ import cls from '@/features/profile/components/PostsBlock/PostsBlock.module.scss
 import {addPostAC, updatePostTexAC} from "@/features/profile/profile-reducer.ts";
 import {Posts} from "@/features/profile/components/PostsBlock/Posts/Posts.tsx";
 import {AddItemForm} from "@/components/AddItemForm/AddItemForm.tsx";
-import {useAppDispatch, useAppSelector} from "@/common/hooks/useAppDispatch.ts";
+import {useAppDispatch} from "@/common/hooks/useAppDispatch.ts";
+import {useAppSelector} from "@/common/hooks/useAppSelector.ts";
 import {CgAddR} from "react-icons/cg";
 import {FC} from "react";
 
@@ -11,7 +12,6 @@ type PostsBlockPropsType = {
 }
 
 export const PostsBlock: FC<PostsBlockPropsType> = ({user}) => {
-
    const dispatch = useAppDispatch();
    // currentText selector will be removed after form lib connecting
    const currentText = useAppSelector(state => state.profile.newPostText);
