+++ src/pages/Users/Users.tsx
@@ -5,14 +5,14 @@ import {getUsersThunkCreator} from "@/features/users/users-thunks.ts";
 import {useUsersData} from "@/features/users/hooks/useUsersData.ts";
 import {withAuthRedirect} from "@/common/hoc/withAuthRedirect.tsx";
 import {setCurrentPage} from "@/features/users/users-reducer.ts";
+import {useAppDispatch} from "@/common/hooks/useAppDispatch.ts";
+import {useAppSelector} from "@/common/hooks/useAppSelector.ts";
 import {User} from "@/features/users/components/User/User.tsx";
-import {useAppDispatch, useAppSelector} from "@/common/hooks/useAppDispatch.ts";
 import {Loading} from "@/components/Loading/Loading.tsx";
 import cls from "./Users.module.scss";
 import {useEffect} from "react";
 
 export const Users = withAuthRedirect(() => {
-
     const {followList, totalUsersCount, currentPage, pageSize, users} = useUsersData();
 
     const appStatus = useAppSelector(appStatusSelector);
